Fix packet buffer OoS, maybe implement size check

+++ HHVMCraft.Core/Networking/Packets/PlayerLookPacket.php
@@ -10,8 +10,8 @@ class PlayerLookPacket {
 	public $onGround;
 
 	public function readPacket($StreamWrapper) {
-		$this->yaw = $StreamWrapper->readUInt8();
-		$this->pitch = $StreamWrapper->readUInt8();
+		$this->yaw = $StreamWrapper->readInt();
+		$this->pitch = $StreamWrapper->readInt();
 		$this->onGround = $StreamWrapper->readBool();
 	}
 }
